Метод для создания плейлиста пересен в PlaylistService
