+++ src/main/java/com/ilnitsk/animusic/playlist/controller/PlaylistController.java
@@ -1,8 +1,14 @@
 package com.ilnitsk.animusic.playlist.controller;
 
+import com.ilnitsk.animusic.playlist.dao.Playlist;
 import com.ilnitsk.animusic.playlist.service.PlaylistService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -11,5 +17,13 @@ import org.springframework.web.bind.annotation.RestController;
 public class PlaylistController {
     private final PlaylistService playlistService;
 
-
+    @PostMapping
+    @Operation(summary = "Метод для создания плейлиста")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Успешное создание"),
+            @ApiResponse(responseCode = "500", description = "Ошибка на стороне сервера")
+    })
+    public Playlist createPlaylist(@RequestParam String playlistName) {
+        return playlistService.createPlaylist(playlistName);
+    }
 }
